+++ src/components/command-palette/index.ts
@@ -1,0 +1,3 @@
+export * from './CommandPalette';
+export * from './commandItems';
+export * from './types';